🐛 Fikset feil med importering.

+++ src/main/java/no/ehfsok/dao/CompanyRepo.java
@@ -1,11 +1,10 @@
 package no.ehfsok.dao;
 
-import java.util.UUID;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 import no.ehfsok.model.Company;
 
 @Repository
-public interface CompanyRepo extends JpaRepository<Company, Integer> {
+public interface CompanyRepo extends JpaRepository<Company, Long> {
 
 }
